refactor: grouped related modules to graph-helper
