+++ frontend/components/common/graph-helper.tsx
@@ -1,0 +1,76 @@
+import { Layout } from "plotly.js";
+import { cloneDeep } from "lodash";
+import { eigs, cos, sin, pi, range, atan2, transpose } from "mathjs";
+
+export const latentGraphLayout = (title: string): Partial<Layout> => {
+  return {
+    title: {
+      text: title,
+    },
+    plot_bgcolor: "#EDEDED",
+    xaxis: {
+      color: "#FFFFFF",
+      tickfont: {
+        color: "#000000",
+      },
+      range: [-3.5, 3.5],
+      gridcolor: "#FFFFFF",
+    },
+    yaxis: {
+      color: "#FFFFFF",
+      tickfont: {
+        color: "#000000",
+      },
+      range: [-3.5, 3.5],
+      gridcolor: "#FFFFFF",
+    },
+    legend: {
+      yanchor: "top",
+      y: 1,
+      x: 0,
+      bgcolor: "rgba(255,255,255,0.8)",
+    },
+    hoverlabel: {
+      font: {
+        family: "monospace",
+      },
+    },
+    clickmode: "event+select",
+    margin: {
+      l: 40,
+      r: 40,
+      b: 40,
+      t: 40,
+      pad: 5,
+    },
+  };
+};
+
+export const calculateGMMRings = (mu: number[], sigma: number[][]) => {
+  sigma = cloneDeep(sigma);
+  const eig = eigs(sigma);
+  let [lambda1, lambda2] = eig.values as number[];
+  let [v1, v2] = transpose(eig.vectors) as number[][];
+
+  if (lambda1 < lambda2) {
+    [lambda1, lambda2] = [lambda2, lambda1];
+    [v1, v2] = [v2, v1];
+  }
+
+  const [v1x, v1y] = v1;
+  const theta = atan2(v1y, v1x);
+  const width = 2 * Math.sqrt(lambda1);
+  const height = 2 * Math.sqrt(lambda2);
+
+  const trace = range(0, 2 * pi, 0.01, true)
+    .map((t) => {
+      const x =
+        mu[0] + width * cos(t) * cos(theta) - height * sin(t) * sin(theta);
+      const y =
+        mu[1] + width * cos(t) * sin(theta) + height * sin(t) * cos(theta);
+      return [x, y];
+    })
+    .toArray() as number[][];
+
+  return trace;
+};
