+++ frontend/components/gmm-home/latent-graph/latent-space-plot.tsx
@@ -1,45 +1,18 @@
-import { cloneDeep, zip } from "lodash";
-import { eigs, cos, sin, pi, range, atan2, transpose } from "mathjs";
+import { zip } from "lodash";
 import dynamic from "next/dynamic";
 import { PlotData } from "plotly.js";
 import { useMemo } from "react";
 import { Card } from "react-bootstrap";
-import { latentGraphLayout } from "~/components/common/graph-layout";
+import {
+  calculateGMMRings,
+  latentGraphLayout,
+} from "~/components/common/graph-helper";
 import { useSelector } from "react-redux";
 import { RootState } from "../redux/store";
 
 import { Props } from ".";
 
 const Plot = dynamic(() => import("react-plotly.js"), { ssr: false });
-
-const calculateTraces = (mu: number[], sigma: number[][]) => {
-  sigma = cloneDeep(sigma);
-  const eig = eigs(sigma);
-  let [lambda1, lambda2] = eig.values as number[];
-  let [v1, v2] = transpose(eig.vectors) as number[][];
-
-  if (lambda1 < lambda2) {
-    [lambda1, lambda2] = [lambda2, lambda1];
-    [v1, v2] = [v2, v1];
-  }
-
-  const [v1x, v1y] = v1;
-  const theta = atan2(v1y, v1x);
-  const width = 2 * Math.sqrt(lambda1);
-  const height = 2 * Math.sqrt(lambda2);
-
-  const trace = range(0, 2 * pi, 0.01, true)
-    .map((t) => {
-      const x =
-        mu[0] + width * cos(t) * cos(theta) - height * sin(t) * sin(theta);
-      const y =
-        mu[1] + width * cos(t) * sin(theta) + height * sin(t) * cos(theta);
-      return [x, y];
-    })
-    .toArray() as number[][];
-
-  return trace;
-};
 
 const useVaeDataPlot = (vaeData: Props["vaeData"], minCount: number) => {
   const vaeDataPlot: Partial<PlotData> = useMemo(() => {
@@ -91,7 +64,7 @@ const useGmmDataPlot = (gmmData: Props["gmmData"]) => {
         "[" + gmmData.means[i].map((d) => d.toFixed(4)).join(", ") + "]";
 
       const trace = zip(
-        ...calculateTraces(gmmData.means[i], gmmData.covariances[i])
+        ...calculateGMMRings(gmmData.means[i], gmmData.covariances[i])
       ) as unknown as number[][];
       const plotData: Partial<PlotData> = {
         name: `MoG No.${i}`,
